Time: 5 ms (19.92%), Space: 7.5 MB (49.73%) - LeetHub

+++ 0397-integer-replacement/0397-integer-replacement.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
-    int solve(int n){
+    int solve(long long n){
         if(n<1)return INT_MAX;
         if(n==1)return 0;
-        int ch1=INT_MAX;
+        long long ch1=INT_MAX;
         if(n%2==0){
             ch1=solve(n/2);
         }else ch1=solve(n-1);
         if(ch1!=INT_MAX)ch1++;
 
-        int ch2=INT_MAX;
+        long long ch2=INT_MAX;
         if(n%2==1)ch2=solve(n+1);
         if(ch2!=INT_MAX)ch2++;
